+++ agent/src/main/java/com/cmclinnovations/agent/utils/LocalisationResource.java
@@ -32,6 +32,8 @@ public class LocalisationResource {
   public static final String SUCCESS_CONTRACT_TASK_REPORT_KEY = SUCCESS_CONTRACT_PREFIX + "task.report";
   public static final String SUCCESS_CONTRACT_TASK_ACCRUAL_KEY = SUCCESS_CONTRACT_PREFIX + "task.accrual";
   public static final String SUCCESS_CONTRACT_TASK_EXEMPT_KEY = SUCCESS_CONTRACT_PREFIX + "task.exempt";
+  public static final String SUCCESS_CONTRACT_TASK_PRIORITY_KEY = SUCCESS_CONTRACT_PREFIX + "task.priority";
+  public static final String SUCCESS_CONTRACT_TASK_PRIORITY_REMOVE_KEY = SUCCESS_CONTRACT_PREFIX + "task.priority.remove";
   public static final String SUCCESS_DELETE_KEY = SUCCESS_PREFIX + "delete";
   public static final String SUCCESS_UPDATE_KEY = SUCCESS_PREFIX + "update";
   public static final String ERROR_ADD_KEY = ERROR_PREFIX + "add";